cortex-bot/requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
ibm-watsonx-ai==1.1.16
